+++ index.html
@@ -230,6 +230,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <script src="js/main.js"></script>
+  <script src="main.js" type="module"></script>
   </body>
 </html>
